Change name of 'current'-screen components
