+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/current/currentpatients/CurrentPatientItem.kt
