+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/current/currentpatients/CurrentPatientsData.kt
@@ -1,7 +1,7 @@
 package com.EENX15_22_17.digital_journal.android.ui.current.currentpatients
 
 // Temporary dataclass before backend is implemented
-data class PatientsData(
+data class CurrentPatientsData(
     val name: String,
     val id: String,
     val securityNumber: String
