--- androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/ui/currentpatients/CurrentPatientsData.kt
@@ -1,12 +1,0 @@
-package com.EENX15_22_17.digital_journal.android.ui.currentpatients
-
-// Temporary dataclass before backend is implemented
-data class CurrentPatientsData(
-    val name: String,
-    val id: String,
-    val securityNumber: String
-)
-
-
-
-
